Allow for richef property assertion messages
issues: closes #6044

Previously, Pproperty.asserts could be provided a fucntion accepting no
arguments to call when an assertion failed. This PR modifies he Property
machinery to expect and call a function accepting (obj, name, value) so
that error messages can be more informative.

This PR also adds a new core/property/tests/test_bases.py to maintain the
property assertion functionality.

Finally, this PR removes an unused code path for assertions that would
automatically raise a ValueError if the assertion function returned a
value other than True or False, on the grounds that explicitly
separating th check and message is preferable, and to simplify the
logic. A (python) assertion was added to enfore that a (Bokeh) assertion
returns a bool, always.